refactor: move methods for attr events
